Improve backtraces configuration.
The previous situation was confusing with various contradicting
(staments of) defaults, mixed with old style and new style
configuration.

Simplify by deciding that having backtraces is so
useful that it is the defaut, both old and new style configurations.
(It was already default on in the config template,
which seems 'canonical'.)

(cherry picked from commit 61453e964cb9223626b4232729674936443d2e9d)